restructuring
Deleted some unused files and tried to clean up the repo
