--- src/Activity.jsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import Layout from './Layout.jsx'
-import ActivityForm from './components/ActivityForm.jsx'
-
-const Activity = () => {
-
-    return (
-        <Layout>
-            <ActivityForm />
-        </Layout>
-    )
-}
-
-export default Activity
