--- src/Form.css
@@ -1,28 +1,0 @@
-*{
-    background: white;
-}
-.row-1{
-    display: flex;
-    justify-content: center;
-    width: 100vw;
-    justify-content: space-evenly;
-}
-
-.form-group{
-   margin-bottom: 20px;
-}
-
-.left-form-activity{
-    margin-bottom: 20px;
-    
-}
-.row-2{
-    display: flex;
-}
-.submit{
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-}
-
-
