--- src/Home.jsx
@@ -1,14 +1,0 @@
-import './App.css';
-import Layout from './Layout.jsx';
-import ActivityList from './components/ActivityList';
-
-const Home = () => {
-
-  return (
-    <Layout>
-      <ActivityList />
-    </Layout>
-  )
-}
-
-export default Home
