+++ src/pages/Home.jsx
@@ -1,0 +1,13 @@
+import Layout from '../Layout.jsx';
+import ActivityList from '../components/ActivityList.jsx';
+
+const Home = () => {
+
+  return (
+    <Layout>
+      <ActivityList/>
+    </Layout>
+  )
+}
+
+export default Home
