+++ src/pages/Profile.jsx
@@ -1,6 +1,5 @@
 import React from "react";
-import Layout from "./Layout.jsx";
-import './App.css'
+import Layout from "../Layout.jsx";
 
 const Profile = () => {
   return (
